learned not mess with interfaces you dont understamd

+++ commands/commands.go
@@ -343,7 +343,7 @@ func quote(ctx context.Context, s *discordgo.Session, m *discordgo.MessageCreate
 		// quote takes into consideration nubmer of reacts for considering which message to quote, more reactions means higher chance
 		// messageDefaultWeight is relative to weight of a single reaction
 		weightsSlice := make([]uint, messages.Len())
-		messageDefaultWeight := viper.Get("bot.quote.default_message_weight").(uint)
+		messageDefaultWeight := uint(viper.Get("bot.quote.default_message_weight").(int))
 		var totalWeight uint = 0
 		// duplicating code inside for loop
 		for i := range weightsSlice {
